requirements.txt
requests
python-dotenv
click
jinja2
pytest
